% Auto-generated: do not edit by hand
\name{toolBox}

\alias{toolBox}

\title{ToolBox component}

\description{

}

\usage{
toolBox(breakpoints=NULL, layouts=NULL, toolboxComponent=NULL,
toolboxItems=NULL, toolboxTitle=NULL)
}

\arguments{
\item{breakpoints}{Named list. }

\item{layouts}{Named list. }

\item{toolboxComponent}{}

\item{toolboxItems}{Unnamed list. }

\item{toolboxTitle}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

