% Auto-generated: do not edit by hand
\name{toolbox}

\alias{toolbox}

\title{Toolbox component}

\description{

}

\usage{
toolbox(breakpoints=NULL, layouts=NULL, toolboxComponent=NULL,
toolboxItems=NULL, toolboxTitle=NULL)
}

\arguments{
\item{breakpoints}{Named list. }

\item{layouts}{Named list. }

\item{toolboxComponent}{}

\item{toolboxItems}{Unnamed list. }

\item{toolboxTitle}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

